src/modules/Counter/model/slice/counterSlice/counterSlice.ts
import { createSlice } from '@reduxjs/toolkit';
import type { PayloadAction } from '@reduxjs/toolkit';

import { CounterSchema } from '../../types/counterSchema';

const initialState: CounterSchema = { value: 0 };

const counterSlice = createSlice({
  name: 'counter',
  initialState,
  reducers: {
    increment(state) {
      state.value++;
    },
    decrement(state) {
      state.value--;
    },
    incrementByAmount(state, action: PayloadAction<number>) {
      state.value += action.payload;
    }
  }
});

export const {
  reducer: counterReducer,
  actions: counterActions
} = counterSlice;
